refactor: Custom Error, Room
