+++ bookings/services/service_v1_booking.py
@@ -4,9 +4,8 @@ from django.utils import timezone
 
 from bookings.models import Booking
 from rooms.models.room import Room
-
 from users.models import User
-from utils.choices import BookingKindChoices
+from utils.choices.choice import BookingKindChoices
 from utils.exceptions.exception import BookingExceptions
