+++ utils/exceptions/handler.py
@@ -1,14 +1,14 @@
 from __future__ import annotations
+
 import logging
 from typing import Any, Optional
+
 from django.core.exceptions import ValidationError as DjangoValidationError
 from django.http import Http404
 from rest_framework import exceptions, serializers
-from rest_framework.exceptions import ErrorDetail
 from rest_framework.response import Response
 from rest_framework.serializers import as_serializer_error
 from rest_framework.views import exception_handler
-
 
 VALIDATION_ERROR_CODE = 999999
 
@@ -40,19 +40,18 @@ def custom_exception_handler(exc: Any, context: Any) -> Optional[Response]:
                 logging.error("[Exception] - List exception, Handling")
                 data["error_detail"] = response.data[0]
             elif isinstance(response.data, dict):
-                logging.error("[Exception] - Dict exception Handling")
-
                 targets = [
                     {"string": key, "code": val} for key, val in response.data.items()
                 ]
             data["targets"] = targets
             response.data = data
+            logging.error(f"[Exception] - ValidationError {exc} {context}")
 
         else:
             # custom exception
             error_detail = response.data.pop("detail", None)
             targets = [
-                {"string": str(error_detail), "code": error_detail.code}
+                {"string": str(error_detail), "code": error_detail.code}  # type: ignore
                 if error_detail is not None
                 else "N/A"
             ]
